# app/backend/static/swagger_api.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "Blueprint",
    "version" : "3.0",
    "title" : "Swagger - Blueprint API",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "email" : "someone@somewhere.com"
    }
  },
  "tags" : [ {
    "name" : "blueprint",
    "description" : "operations on blueprint"
  }, {
    "name" : "user",
    "description" : "operations on user"
  }, {
    "name" : "template",
    "description" : "operations on template"
  } ],
  "paths" : {
    "/api/blueprint-templates" : {
      "get" : {
        "tags" : [ "template" ],
        "summary" : "get all templates",
        "operationId" : "listTemplates",
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "post" : {
        "tags" : [ "template" ],
        "summary" : "create template",
        "operationId" : "createTemplate",
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/blueprint-template"
              }
            }
          },
          "description" : "create blueprint template object",
          "required" : true
        }
      }
    },
    "/api/blueprint-templates/{template_id}" : {
      "get" : {
        "tags" : [ "template" ],
        "summary" : "get template by id",
        "operationId" : "getTemplate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "the template to get",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/blueprint-template"
                }
              }
            }
          },
          "400" : {
            "description" : "invalid template id supplied"
          },
          "404" : {
            "description" : "template not found"
          }
        }
      },
      "put" : {
        "tags" : [ "template" ],
        "summary" : "updated template",
        "operationId" : "updateTemplate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id of template to update",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "invalid template supplied"
          },
          "404" : {
            "description" : "template not found"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/blueprint-template"
              }
            }
          },
          "description" : "updated template object",
          "required" : true
        }
      },
      "delete" : {
        "tags" : [ "template" ],
        "summary" : "delete template",
        "operationId" : "deleteTemplate",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id of template to delete",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "invalid template supplied"
          },
          "404" : {
            "description" : "template not found"
          }
        }
      }
    },
    "/api/blueprints" : {
      "get" : {
        "tags" : [ "blueprint" ],
        "summary" : "get all blueprints",
        "operationId" : "listBlueprints",
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "post" : {
        "tags" : [ "blueprint" ],
        "summary" : "create blueprint",
        "operationId" : "createBlueprint",
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/blueprint"
              }
            }
          },
          "description" : "create blueprint object",
          "required" : true
        }
      }
    },
    "/api/blueprints/{blueprint_id}" : {
      "get" : {
        "tags" : [ "blueprint" ],
        "summary" : "get blueprint by id",
        "operationId" : "getBlueprint",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "the blueprint to get",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/blueprint"
                }
              }
            }
          },
          "400" : {
            "description" : "invalid blueprint name supplied"
          },
          "404" : {
            "description" : "blueprint not found"
          }
        }
      },
      "put" : {
        "tags" : [ "blueprint" ],
        "summary" : "updated blueprint",
        "operationId" : "updateBlueprint",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id of blueprint to update",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "invalid blueprint supplied"
          },
          "404" : {
            "description" : "blueprint not found"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/blueprint"
              }
            }
          },
          "description" : "updated user object",
          "required" : true
        }
      },
      "delete" : {
        "tags" : [ "blueprint" ],
        "summary" : "delete blueprint",
        "operationId" : "deleteBlueprint",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id of blueprint to delete",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "invalid blueprint supplied"
          },
          "404" : {
            "description" : "blueprint not found"
          }
        }
      }
    },
    "/api/users" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "get all users",
        "operationId" : "listUsers",
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "post" : {
        "tags" : [ "user" ],
        "summary" : "create user",
        "operationId" : "createUser",
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/user"
              }
            }
          },
          "description" : "create user object",
          "required" : true
        }
      }
    },
    "/api/users/{username}" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "get user by username",
        "operationId" : "getUser",
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "description" : "the user to get",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/user"
                }
              }
            }
          },
          "400" : {
            "description" : "invalid username supplied"
          },
          "404" : {
            "description" : "user not found"
          }
        }
      },
      "put" : {
        "tags" : [ "user" ],
        "summary" : "updated user",
        "operationId" : "updateUser",
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "description" : "name of the user to update",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "invalid user supplied"
          },
          "404" : {
            "description" : "user not found"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/user"
              }
            }
          },
          "description" : "updated user object",
          "required" : true
        }
      },
      "delete" : {
        "tags" : [ "user" ],
        "summary" : "delete user",
        "operationId" : "deleteUser",
        "parameters" : [ {
          "name" : "username",
          "in" : "path",
          "description" : "the name of the user to delete",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "invalid username supplied"
          },
          "404" : {
            "description" : "user not found"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "user" : {
        "type" : "object",
        "required" : [ "id", "username" ],
        "properties" : {
          "username" : {
            "type" : "string",
            "description" : "soe id"
          },
          "email" : {
            "type" : "string",
            "description" : "email"
          },
          "cmpRequests" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "properties" : {
                "cmp" : {
                  "type" : "string",
                  "description" : "cmp request id, tentative until we know what abstraction tool provides"
                },
                "status" : {
                  "type" : "string",
                  "enum" : [ "pending", "completed", "failed" ]
                }
              }
            }
          },
          "status" : {
            "type" : "string",
            "description" : "user status",
            "enum" : [ "pending", "active", "inactive" ]
          }
        }
      },
      "blueprint" : {
        "type" : "object",
        "required" : [ "id", "name", "cluster", "user" ],
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "cluster" : {
            "type" : "string",
            "description" : "openshift cluster where blueprint is deployed"
          },
          "user" : {
            "type" : "string",
            "description" : "user owner of this blueprint"
          },
          "createDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "exempt" : {
            "type" : "boolean",
            "description" : "exempt blueprint from auto delete after X days"
          },
          "region" : {
            "type" : "string",
            "description" : "region/datacenter where blueprint is deployed",
            "enum" : [ "NAM", "LATAM", "APAC", "EU" ]
          },
          "status" : {
            "type" : "string",
            "description" : "status of the blueprint",
            "enum" : [ "pending", "active", "deleted" ]
          }
        }
      },
      "blueprint-template" : {
        "type" : "object",
        "required" : [ "id", "author", "num_likes" ],
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "author" : {
            "type" : "string"
          },
          "num_likes" : {
            "type" : "string",
            "description" : "number of likes on a template"
          },
          "liked_by_users" : {
            "type" : "array",
            "description" : "list of users that liked this template",
            "items" : {
              "type" : "string",
              "description": "user"
            }
          }
        }
      }
    }
  }
}